Possible fix for alarm not triggered bug
On some devices, the alarm was not being triggered at the correct time, or the alarm services were being killed in the middle. This commit contains a possible fix: all the PendingIntents are now created with no flags (0), and all alarm-related Intents and PendingIntents are created with application context. Not sure if this is "the" fix.