Removed support for Intent.ACTION_DATE_CHANGED
The MyApplication class was being used to notify Activity_AlarmsList when Intent.ACTION_DATE_CHANGED was broadcast by the system. Extensive tests, and one crash report from Google Play, showed that this was not working as intended as the logic was somewhat flawed. We have completely removed support for this broadcast as of now.