Manage new alarm fired when another is snoozed
Suppose an alarm is currently snoozed. In the mean time, another alarm fires (Service_RingAlarm starts). In this situation, Service_RingAlarm will check whether its alarmID is the same as that of Service_SnoozeAlarm. If different, it will ask the latter to stop. Due to #27, Service_SnoozeAlarm will dismiss the alarm before being destroyed.