Fix snooze not working bug
On many devices, there is a problem that even though Service_SnoozeAlarm is started successfully and it runs without errors, the countdown does not complete at the specified interval. The problem is completely erratic — if the snooze interval is 5 mins, sometimes even the first snooze does not work; sometimes the second snooze is fired a long time after the first one. But the service keeps on running. The probable reason is that the CPU goes into doze mode when the screen is turned off, so countdown stops. The solution is to hold a wakelock throughout the lifetime of this service. The fix has been tested to work on Marshmallow and Q.