Modified handling of permission in RingtonePicker
When Activity_RingtonePicker, it will first check whether READ_EXTERNAL_STORAGE permission is available. If available, then it will load the tones. If not available, then it will first ask for the permission. When granted, the tones will be loaded, otherwise the activiy will finish off with a Toast message.