Check non-ess. perms in UI thread in AlarmsList
The check for non-essential perms will now occur in the UI thread instead of background thread in Activity_AlarmsList. Reason: MutableLiveData.setValue() does not work in background threads, and we have to use postValue(). As a result, the value is actually not being set even after the thread finishes, and by that time, the activity is passing onResume(). If the app ever pauses and onResume() is called again, the perms. activity is suddenly launched. Hence, it is necessary to check for the non-essential perms in the main thread itself.